Build test SDE resources.
