+++ src/main/java/com/autonomouslogic/everef/cli/CommandRunner.java
@@ -46,6 +46,9 @@ public class CommandRunner {
 	protected Provider<ScrapeHoboleaks> scrapeHoboleaksProvider;
 
 	@Inject
+	protected Provider<ImportTestResources> importTestResourcesProvider;
+
+	@Inject
 	protected HealthcheckDecorator healthcheckDecorator;
 
 	@Inject
@@ -98,6 +101,8 @@ public class CommandRunner {
 				return verifyRefDataModelsProvider.get();
 			case "scrape-hoboleaks":
 				return scrapeHoboleaksProvider.get();
+			case "import-test-resources":
+				return importTestResourcesProvider.get();
 			default:
 				throw new IllegalArgumentException("Unknown command: " + name);
 		}
