+++ src/test/resources/refdata/sde/fsd/categoryIDs.yaml
@@ -1,10 +1,11 @@
-6:
+---
+"6":
   name:
-    de: Schiff
-    en: Ship
-    es: Nave
-    fr: Vaisseau
-    ja: 艦船
-    ru: Корабль
-    zh: 舰船
+    de: "Schiff"
+    en: "Ship"
+    es: "Nave"
+    fr: "Vaisseau"
+    ja: "艦船"
+    ru: "Корабль"
+    zh: "舰船"
   published: true
