+++ src/test/resources/refdata/sde/fsd/groupIDs.yaml
@@ -1,30 +1,31 @@
-1185:
+---
+"1185":
   anchorable: false
   anchored: false
   categoryID: 11
   fittableNonSingleton: false
   name:
-    de: FW Schlachtkreuzer der Gallente-Föderation
-    en: FW Gallente Federation Battlecruiser
-    es: Crucero de combate de GF de la Federación Gallente
-    fr: Croiseur de bataille Fédération gallente GF
-    ja: FWガレンテ連邦巡洋戦艦
-    ru: Линейный крейсер ополчения Галлентской Федерации
-    zh: 势力战争盖伦特联邦战列舰
+    de: "FW Schlachtkreuzer der Gallente-Föderation"
+    en: "FW Gallente Federation Battlecruiser"
+    es: "Crucero de combate de GF de la Federación Gallente"
+    fr: "Croiseur de bataille Fédération gallente GF"
+    ja: "FWガレンテ連邦巡洋戦艦"
+    ru: "Линейный крейсер ополчения Галлентской Федерации"
+    zh: "势力战争盖伦特联邦战列舰"
   published: true
   useBasePrice: false
-257:
+"257":
   anchorable: false
   anchored: false
   categoryID: 16
   fittableNonSingleton: false
   name:
-    de: Raumschiff-Kommando
-    en: Spaceship Command
-    es: Manejo de naves espaciales
-    fr: Pilotage de vaisseau
-    ja: 宇宙船操作
-    ru: Допуски к управлению кораблями
-    zh: 飞船操控学
+    de: "Raumschiff-Kommando"
+    en: "Spaceship Command"
+    es: "Manejo de naves espaciales"
+    fr: "Pilotage de vaisseau"
+    ja: "宇宙船操作"
+    ru: "Допуски к управлению кораблями"
+    zh: "飞船操控学"
   published: true
   useBasePrice: true
