+++ src/test/resources/refdata/sde/fsd/marketGroups.yaml
@@ -1,28 +1,29 @@
-1376:
+---
+"1376":
   descriptionID:
-    de: Große Raumschiffe mit ausreichend offensiven und defensiven Möglichkeiten,
-      um alle Probleme - bis auf die größten - in Agriff zu nehmen
-    en: Large-sized spaceships with enough offensive and defensive capabilities to
-      tackle all but the biggest problems
-    es: Naves de tamaño grande con suficientes capacidades ofensivas y defensivas
-      para enfrentarse a cualquier problema... excepto a los más gordos.
-    fr: Vaisseaux spatiaux de grande taille disposant de suffisamment de capacités
-      offensives et défensives pour s'attaquer à toutes les cibles hormis les plus
-      grandes
-    ja: 大抵の敵艦に立ち向かえるだけの攻撃力・防御力を誇る大型戦艦
-    ko: 강력한 공격 및 방어 성능을 보유한 대형 함선으로 함대의 주력 함선입니다.
-    ru: Корабли большого тоннажа, обладающие достаточными защитой и вооружением, чтобы
-      справиться даже с самыми крупными неприятностями
-    zh: 大型舰船，攻防能力出色，足以应对大多数局面
+    de: "Große Raumschiffe mit ausreichend offensiven und defensiven Möglichkeiten,\
+      \ um alle Probleme - bis auf die größten - in Agriff zu nehmen"
+    en: "Large-sized spaceships with enough offensive and defensive capabilities to\
+      \ tackle all but the biggest problems"
+    es: "Naves de tamaño grande con suficientes capacidades ofensivas y defensivas\
+      \ para enfrentarse a cualquier problema... excepto a los más gordos."
+    fr: "Vaisseaux spatiaux de grande taille disposant de suffisamment de capacité\
+      s offensives et défensives pour s'attaquer à toutes les cibles hormis les plus\
+      \ grandes"
+    ja: "大抵の敵艦に立ち向かえるだけの攻撃力・防御力を誇る大型戦艦"
+    ko: "강력한 공격 및 방어 성능을 보유한 대형 함선으로 함대의 주력 함선입니다."
+    ru: "Корабли большого тоннажа, обладающие достаточными защитой и вооружением,\
+      \ чтобы справиться даже с самыми крупными неприятностями"
+    zh: "大型舰船，攻防能力出色，足以应对大多数局面"
   hasTypes: false
   iconID: 1443
   nameID:
-    de: Schlachtschiffe
-    en: Battleships
-    es: Acorazados
-    fr: Cuirassés
-    ja: 戦艦
-    ko: 배틀쉽
-    ru: Линкоры
-    zh: 战列舰
+    de: "Schlachtschiffe"
+    en: "Battleships"
+    es: "Acorazados"
+    fr: "Cuirassés"
+    ja: "戦艦"
+    ko: "배틀쉽"
+    ru: "Линкоры"
+    zh: "战列舰"
   parentGroupID: 4
